feat: 프로필, 로고 이미지 초기화하는 API 추가

+++ backend/src/projects/projects.service.ts
@@ -180,10 +180,12 @@ export class ProjectsService {
     } else if (user.id !== project.ownerId) {
       throw new ForbiddenException('👻 프로젝트 등록자만 변경할 수 있어요 🌫');
     } else {
-      const key = project.logoImageUrl.split(
-        `https://${process.env.AWS_S3_BUCKET_NAME}.s3.amazonaws.com/`,
-      )[1];
-      await this.awsService.deleteS3Object(key);
+      if (project.logoImageUrl) {
+        const key = project.logoImageUrl.split(
+          `https://${process.env.AWS_S3_BUCKET_NAME}.s3.amazonaws.com/`,
+        )[1];
+        await this.awsService.deleteS3Object(key);
+      }
 
       const s3Object = await this.awsService.uploadFileToS3(
         'projects',
@@ -192,6 +194,29 @@ export class ProjectsService {
       const logoImageUrl = this.awsService.getAwsS3FileUrl(s3Object.key);
       await this.projectsRepository.update(id, {
         logoImageUrl,
+      });
+
+      return await this.projectsRepository.find({ where: { id } });
+    }
+  }
+
+  async initLogoImage(user, id) {
+    const project = await this.projectsRepository.findOne({
+      select: ['ownerId', 'logoImageUrl'],
+      where: { id, deletedAt: null },
+    });
+    if (!project) {
+      throw new NotFoundException('👻 존재하지 않는 프로젝트에요 🌫');
+    } else if (user.id !== project.ownerId) {
+      throw new ForbiddenException('👻 프로젝트 등록자만 변경할 수 있어요 🌫');
+    } else {
+      const key = project.logoImageUrl.split(
+        `https://${process.env.AWS_S3_BUCKET_NAME}.s3.amazonaws.com/`,
+      )[1];
+      await this.awsService.deleteS3Object(key);
+
+      await this.projectsRepository.update(id, {
+        logoImageUrl: null,
       });
 
       return await this.projectsRepository.find({ where: { id } });
